Quantity format option to display units in abbreviated form
If the format spec ends in ~, the abbreviated form of the unit name
will be used. For example:

    >>> q = Q_(1, 'kilometer'),
    >>> print('{!s}'.format(q))
    1 kilometer
    >>> print('{!s~}'.format(q))
    1 km

The same is true for the repr (r), latex (l) and pretty (f) format.
The abbreviated form is specified by the first alias in the definition
file. If no abbreviation is found, the long name is used.