Deduplicate units from candidates search
Previous to this commit, _parse_candidate was returning an iterable of
possible units, including prefixes and suffixes combinations.
If multiple values where returned, the calling functions were issuing
a warning to inform the user that there are multiple matches for a
given string. For example, if a user defines a new unit ms, it could
be mistaken with prefix=mili, unit=second. However, as the registry
stores also prefixed units many times a duplication is not a real
problem (eg. prefix=kilo, unit=gram and unit=kilogram).
To avoid issuing false warnings a deduplication step was added to
parse candidates. In the future, a more efficient alternative will be
to block adding colliding units to the registry.