Parsing contexts from a string.
A context definition starts with `@context` followed by
the name. Relationships between dimensions are specified
one per line. Left hand side of the equal sign is used to
indicate which dimensions are related. A single arrow (`->`)
is used to indicate a unidirectional relationship. A double
arrow (`<->`) is used to indicate a bidirectional relationship

To the right, the equation is given using value as the variable name.

For example:

@context spectral
    [time] <-> 1 / [length] = c / value

Parameters and aliases can be given in the header.

@context(n=1) spectra = sp

Still to be done:
- a proper API to define contexts in the registry.
- the way to signal end of a context in the defintion file.