Store the magnitude when converting to ndarray
When an `__array_` attribute is wanted, it is very likely that was
requested by a NumPy function. We need to convert the magnitude
to an ndarray to obtain the correct, units aware, result.

Previously, the magnitude value was converted to a ndarray and returned,
but the ndarray was not stored in the `_magnitude` attribute.
Very likely, this lead to some unwanted garbage collection.

See #92.