Make the Default Unit Registry a Lazy object
This commit creates a helper class to allow lazy initializatio of a
UnitRegistry. Instead of adding a 'is_initialized' guard to each method,
we overwrite the __class__ attribute and we initialized the Registry when
an attribute is requested.

Warning: Only methods needed to instantiate lazily a Registry are written.

Pending to see: Portability?

See #108