Implemented in place conversion in the registry.
The converters methods `from_reference` and `to_reference` now take
an optional argument (named `inplace` and defaulting to False) to
indicate if the conversion should be done in place or not.

`UnitRegistry.convert` takes the same argument.

This is convenient when operating with big numpy arrays and do not
want to allocate extra memory.

Note: Context mediated transformations are not done in place yet.