Added new configurable mul/div-operation for offset quantities
pint now refuses by default multiplication/division with offset units (which
are called non-multiplicative for a reason). If this is not desired the user
can override this strict behavior with the new unit.Registry-option
"autoconvert_offset_to_baseunit": If True converts offset units in quantities
to their base units in multiplicative context. Test were made compatible with
both options. New tests for the autoconvert-option were added. test86c was
changed to be an expected failure since it contradict the new rules.