Added support for siunitx formatting.
Added support for generating siunitx commands in the __format__
functions for Quantity, Measurment, and Unit classes.

This adds a new code, 'Lx', to the formatting spec, that if present will
trigger the LaTeX siunitx formatting. All other formatting options in
the spec string are be passed through.

Currently does not accept siunit options.