Implements infrastructure for doctesting Pint
The most important aspect is a custom OutputChecker used to circumvent
certain aspects of an easy to read documentation. It performs a series
of conversions to the actual and expected output string of each doctest.
If the standard checker fails, the following steps are taken stopping
when the first is successful:

  1. eval the strings and compare the results.
     (this is useful to compare unordered collections such as dicts)
  2. parse the repr of a Quantity object and compare values and units
     Importantly, values are compared allowing rounding errors (0.1% difference)
  3. parse the str repr of a Quantity object and compare values and units
     Again, values are compared allowing rounding errors (0.1% difference)
  4. Replace the repr of a Unit by the corresponding string and compare
     the resulting strings.
  5. If at least one replacement was done in 4, then start in 1 again.