More readable `to_compact()`
It is largely a matter of taste but I find "0.1 millimeter" easier to read than "100 micrometer". It is even more blatant in higher dimensions:

100000 nanometer^2 == 0.1 micrometer^2,

pint currently compacts "1e-13 meter" into the former. (The implementation of `to_compact()` enforces a magnitude `> 0`.) With this patch one gets the latter. The tests are updated accordingly.