Remove pandas_array import from __init__.py
This should mean that all the tests pass again.

Given pint's focus on avoiding dependencies, we shouldn't import the pandas interface by default (as it depends on pandas). If users want it, they'll have to do `from pint.pandas_array import QuantityArray` which is a little longer than `from pint import QuantityArray` but I think that's ok as it's a specialised usage.