New Implementation of 'Lx'
This approach uses the default/Python format from uncertainties and then
adds the required fixups to make it compatible with SIunitx input
format.  I think this is the most reasonable option available now, as it
is workable and preserves access to most of the existing type conversion
characters implemented by uncertainties package, but not 'S' or 'P'.