Empty Default SIunitx Option
The default option in 'pint' will override SIunitx settings from the
user's document.  There still should be a way of setting it, but it
should probably be empty to start, since it is easier to change the
document settings than to change 'pint' output.  Also, this conforms to
the example in the tutorial.