Replace 1.0 and 0.0 (floats) by 1 and 0 (integers)
In the original codebase, many accumulators, default values and so on were initialized
with floats (instead fo integers). The reason behind this decision was that in Python 2
the division (/) was by default an integer division if all values were ints. Therefore,
using a float ensured "the right result".

Now that the codebase is Python 3, we can safely use integers everywhere. This wil be
helpful when a different non integer type is desired.