Implements Logaritmic Units
Dimensionless Logaritmic units, such as dB, decade, octave or semitone are implemented.

B, used originally as alias for byte, is now used as alias for bell.
This modification allows dB to be defined using d as prefix.

decade, used originaly as alias for ten years, is now used alias for order of magnitude. This modification allows to define transistor characteristics, such as subthreshold swings in mV/decade.