Implements a first benchmark suite.
The benchmark is based on airspeed velocity (asv). Benchmarks are organized with double digits numbers
to create a nice, sorted

00-09: Benchmarks that do not require a registry created during setup
10-19: Benchmarks calling functions directly from the registry
20-29: Benchmarks involving Units and Quantity objects (with scalar values)
30-39: Benchmarks involving Quantity objects (with array values)

Briefly, you can run the the benchmark calling:

$ asv run

or if you want to compare certain commits/branches:

$ asv run master..mybranch

For convenience, we include a list of hashes of all pint versions:

$ asv run HASHFILE:benchmarks/hashes.txt

--

After running the benchmark, run:

$ asv publish
$ asv preview

To get a nice readable output.

More information on: https://asv.readthedocs.io/en/stable/