Fix to_compact, infer_base_unit for non-float non_int_type (#1461)
to_compact used math.log10, which converted non-float types to floats;
this caused exceptions when using a non-integer type, like Decimal,
that does not implement multiplication/division with floats, and
generally changed non-integer types to floats.  This addresses the
problem by checking non_int_type.  If non_int_type != float, it
assumes that the type implements a log10 method, as Decimal does.

infer_base_unit created a generic UnitsContainer, not a UnitsContainer
specific to a registry; this allows specification of a registry, or
uses the registry of the passed-in Quantity/UnitLike.