Use float for to_compact SI power calculation
For more general non-float non_int_type support, convert all
non_int_type values to floats for SI power calculation in
to_compact. As the value is rounded to an integer, and is used only
for determining compact unit representations, this seems reasonable to
support the widest range of non-integer types.  The previous method
worked for Decimal, but not Fraction, and Fraction does not have a
non-float log10 implementation.

This also adds tests for Fraction to_compact and infer_base_unit.