Implemented new definition parser and rewrite Registry.load_definition to use it
This commit implements a Parser class that is able to read a definition file
generating a DefinitionFile object.

The DefinitionFile object contains information about the source file
(path, modification time and a flag that specifies if it is a resource)
and a tuple with the parsed lines. Each element contains the line number
and a definition object.

Importantly, this DefinitionFile is pickable.

In spite of the additional objects, there is no significant change
in the benchmark.