Implements cache, and add caching capability to the parser
Files are cached with pickle to a folder with a name defined by hashing the fullpath.

If the cache is enabled (now is not), the cached files are used if their m_time
is more recent that the required file m_time. If this is not the case,
a cache file is generated.

Implemented this required some chanegs to the parser, namely reorganizing the
loading of package resources (which might not be a real file and therefore is
not clear how to determine if it was modified).

Parsing and loading definitions is now 5 times faster.