Remove UnitsContainer hash from pickle state
Starting from Python 3.3, the hash value of an object has a random salt
that is chosen at interpreter startup. Therefore, pickling and recovering
and object hash is incorrect if the the dumpeed value will be used at
a different interpreter run.

Until this commit, a UnitsContainer that was pickled and then unpickled
in a different interpreter would be wrongly consider as unequal.

BREAKING CHANGE: This commit will break existing code that relies on
keeping the hash across interpreters run. This should be very rare
as will require to set the PYTHONHASHSEED value.