Add configurable cache for registry
The user can configure the cache folder upon registry instantiation.
If the cache_folder is None, the cache is disabled

The cache is used during definition file parsing and cache building.

In the future, a special value will be used to instruct pint to
automatically select a folder.

Until more testing is done, the cache is disabled by default.