fix LaTeX siunitx exponent formatting (closes #1515)
With siunitx, you can format a quantity in scientfic notation by
specifying the quantity's magnitude in 'e' notation.

\SI{1e2}{\meter}

Here, the '1e2' will expand to '1 \times 10^{2}'.

So, when Pint formats exponents for LaTeX, it should *not* expand '1e2'
to '1 \times 10^{2}' *if* siunitx is used.

This commit adds a check to disable expaning exponents for quantity
magnitude formatting if siunitx is used.