Update error URLs for log / offset units.
Log / offset unit documentation changed location, making the URLs shown
in  error message no longer valid.  In addition, the two are on separate
pages in the most recent documentation.  This commit updates the URLs.
It also, as it seems most likely to be stably useful for users, points
to the latest stable documentation, rather than the latest dev version.