Set cache from read cache files
When building a unit registry using a cache, the read cache data is not
getting leveraged. Instead a new cache is getting built, but without the
dimensional equivalents.

This creates a behavior where you create a new UnitRegistry from already
cached files and call `get_compatible_units` it will always fail with
a key error.