Large commit to make Pint more typing friendly
In this very large commit we tackle a few aspects of Pint that
makes it difficult to do static typing.

1. Dynamic classes became static: Quantity and Unit are now
(for the most part) static classes with a static inheritance.
This allows mypy/pylance and other type checker to properly
inspect them.

2. Added types through out all the code. (WIP)

3. Refactor minor parts of the code to make it more typing
homogeneous. Catch a few potential bugs in the way.

4. Add several TODOs that need to be addressed in 0.23

5. Moved some group and system and context code out of the PlainRegistry

6. Moved certain specialized methods out of the PlainRegistry.