Refactor sort_func into `format_compound_unit`
This is a much more logical place to put it.  Note that the default formatters (plain, html, latex) all now call `formatter` with sort_func=None so that we don't accidentally use `sorted` as a default argument.  But those who call `formatter` directly for their own purposes can call with a sort_func of their choosing that will do what they want it to do.

This also fixes a latent bug where we failed to call `sort_func` in one of the paths of `plain.py`.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>