Skip false xfail tests
In test_compat_downcast.py, 12 tests (from parameter combinations) are marked as xfail, but only 4 of them really fail, and so the other 8 end up appearing as failed tests, which is illogical. To solve this, the 8 false xfail are now skipped. PS: It would better to only mark the 4 real failing tests as xfail, but it is not so easily achieved in the context of parameter combinations.