Fix install and testing issues (closes #2188) (#2197)
* Avoid MIP if Python >= 3.12

The MIP library causes "Windows fatal exception: access violation" since Python version 3.12 (see Issue #2121).
Therefore, we should avoid installing it if Python >= 3.12. This commit implements this behavior. Comments are also added to keep track of the reason.

* Impose Dask < 2025.3.0

Dask causes "RuntimeError: Attempting to use an asynchronous Client in a synchronous context of `dask.compute`" since version 2025.3.0 (see Issue #2188 in Pint and Issue #1016 in Dask). Therefore, this commit imposes Dask < 2025.3.0 at installation. Commens are also added to keep track of the reason.

* Skip false xfail tests

In test_compat_downcast.py, 12 tests (from parameter combinations) are marked as xfail, but only 4 of them really fail, and so the other 8 end up appearing as failed tests, which is illogical. To solve this, the 8 false xfail are now skipped. PS: It would better to only mark the 4 real failing tests as xfail, but it is not so easily achieved in the context of parameter combinations.

* Improve "Contributing" documentation

- Add specific instructions for Windows
- Add requirements of external applications (Graphviz, Pandoc, Pixi)

* Update CHANGES